requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.11.1
blinker==1.4
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
exrex==0.11.0
filelock==3.13.1
Flask==2.1.1
Flask-Cors==3.0.10
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
h8mail==2.5.6
holehe==1.61
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.6
ignorant==1.2
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.3
kamene==0.32
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.5
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
nili @ git+https://github.com/niloofarkheirkhah/nili@285220aba2d9a27c852e7471011d9f6cd7e0daae
numpy==1.26.4
oauthlib==3.2.0
openpyxl==3.0.10
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.36
platformdirs==4.2.0
pycountry==23.12.11
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.1
PyYAML==5.4.1
requests==2.28.1
requests-futures==1.0.1
rfc3986==1.5.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
stem==1.8.2
systemd-python==234
termcolor==2.4.0
torrequest==0.1.0
toutatis==1.26
tqdm==4.66.4
trio==0.25.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.18
virtualenv==20.25.0
wadllib==1.3.6
Werkzeug==3.0.2
wifite==2.2.5
wikipedia==1.4.0
yarl==1.9.4
zipp==1.0.0
